IoT/Wemo/requirements.txt
flask
flask_restful
flask_wtf
ouimeaux
grpcio
grpcio-tools
python-dotenv
pymongo 